+++ tests/Endpoints/UserTest.php
@@ -1,16 +1,26 @@
 <?php
 
-use Gotify\Auth\User as AuthUser;
-use Gotify\Endpoint\User;
+namespace Tests\Endpoint;
 
+use PHPUnit\Framework\Attributes\CoversClass;
+use PHPUnit\Framework\Attributes\UsesClass;
+use Tests\AbstractTestCase;
+use Gotify\Endpoint\User;
+use Gotify\Endpoint\AbstractEndpoint;
+use Gotify\Auth\User as Auth;
+
+#[CoversClass(User::class)]
+#[UsesClass(AbstractEndpoint::class)]
+#[UsesClass(\Gotify\Guzzle::class)]
+#[UsesClass(\Gotify\Json::class)]
+#[UsesClass(\Gotify\Server::class)]
+#[UsesClass(\Gotify\Auth\User::class)]
+#[UsesClass(\Gotify\Auth\AbstractAuth::class)]
 class UserTest extends AbstractTestCase
 {
     private static User $user;
 
     private static int $userId = 0;
-
-    private static string $testUsername = 'test';
-    private static string $testPassword = 'test';
 
     public static function setUpBeforeClass(): void
     {
@@ -20,6 +30,13 @@ class UserTest extends AbstractTestCase
             self::$server,
             self::$auth
         );
+
+        self::removeTestUsers();
+    }
+
+    public static function tearDownAfterClass(): void
+    {
+        self::removeTestUsers();
     }
 
     /**
@@ -35,7 +52,7 @@ class UserTest extends AbstractTestCase
         $this->assertObjectHasProperty('name', $current);
         $this->assertObjectHasProperty('admin', $current);
 
-        $this->assertEquals(self::$username, $current->name);
+        $this->assertEquals('admin', $current->name);
     }
 
     /**
@@ -44,7 +61,6 @@ class UserTest extends AbstractTestCase
     public function testGetUser(): void
     {
         $user = self::$user->getUser(self::$userId);
-
         $this->assertIsObject($user);
     }
 
@@ -69,24 +85,23 @@ class UserTest extends AbstractTestCase
     public function testCreate(): void
     {
         $created = self::$user->create(
-            self::$testUsername,
-            self::$testPassword,
+            'createMe',
+            'qwerty',
             false // Admin status
         );
 
         $this->assertIsObject($created);
-        self::$userId = $created->id;
     }
 
     /**
      * Test updating password for the current user
-     *
-     * @depends testCreate
      */
     public function testUpdatePassword(): void
     {
+        self::$user->create('updateMePassword', 'qwerty');
+
         // Login as test user
-        $auth = new AuthUser(self::$testUsername, self::$testPassword);
+        $auth = new Auth('updateMePassword', 'qwerty');
         $user = new User(
             self::$server,
             $auth
@@ -94,8 +109,7 @@ class UserTest extends AbstractTestCase
 
         $updated = $user->updatePassword('NewPassword');
 
-        $this->assertIsBool($updated);
-        $this->assertEquals(true, $updated);
+        $this->assertTrue($updated);
     }
 
     /**
@@ -106,9 +120,9 @@ class UserTest extends AbstractTestCase
     public function testUpdate(): void
     {
         $newTestUsername = 'test1';
-
+        $created =  self::$user->create('updateMe', 'qwerty');
         $updated = self::$user->update(
-            self::$userId,
+            $created->id,
             $newTestUsername
         );
 
@@ -116,20 +130,30 @@ class UserTest extends AbstractTestCase
         $this->assertObjectHasProperty('name', $updated);
         $this->assertObjectHasProperty('id', $updated);
 
-        $this->assertEquals(self::$userId, $updated->id);
+        $this->assertEquals($created->id, $updated->id);
         $this->assertEquals($newTestUsername, $updated->name);
     }
 
     /**
      * Test deleting a user
-     *
-     * @depends testCreate
      */
     public function testDelete(): void
     {
-        $deleted = self::$user->delete(self::$userId);
+        $created = self::$user->create('deleteMe', 'qwerty');
+        $deleted = self::$user->delete($created->id);
 
         $this->assertIsBool($deleted);
         $this->assertEquals(true, $deleted);
+    }
+
+    private static function removeTestUsers(): void
+    {
+        $users = self::$user->getAll();
+
+        foreach ($users->users as $user) {
+            if ($user->admin === false) {
+                self::$user->delete($user->id);
+            }
+        }
     }
 }