+++ CHANGELOG.md
@@ -2,6 +2,23 @@
 
 All notable changes to this project are documented in this file.
 
+## [1.9.0](https://github.com/VerifiedJoseph/gotify-api-php/releases/tag/v1.9.0) - 2024-08-01
+
+- Renamed abstract classes ([#399](https://github.com/VerifiedJoseph/gotify-api-php/pull/399), [`62f2bcb`](https://github.com/VerifiedJoseph/gotify-api-php/commit/62f2bcbddd1834ef31a932050f7e8e79d8a1ee27))
+- Updated dependency `guzzlehttp/guzzle` from 7.9.1 to 7.9.2 ([#402](https://github.com/VerifiedJoseph/gotify-api-php/pull/402), [`476d619`](https://github.com/VerifiedJoseph/gotify-api-php/commit/476d619d647660d79ef15606823b704e304e7ec0))
+
+## [1.8.4](https://github.com/VerifiedJoseph/gotify-api-php/releases/tag/v1.8.4) - 2024-07-20
+
+* Updated dependency `guzzlehttp/guzzle` from 7.9.0 to 7.9.1 ([#396](https://github.com/VerifiedJoseph/gotify-api-php/pull/396), [`dfe3ed4`](https://github.com/VerifiedJoseph/gotify-api-php/commit/dfe3ed452a9fe0e82ca8e2eb38fc028aae6df623))
+
+## [1.8.3](https://github.com/VerifiedJoseph/gotify-api-php/releases/tag/v1.8.3) - 2024-07-18
+
+* Updated dependency `guzzlehttp/guzzle` from 7.8.1 to 7.9.0 ([#394](https://github.com/VerifiedJoseph/gotify-api-php/pull/394), [`9579481`](https://github.com/VerifiedJoseph/gotify-api-php/commit/95794811418b00effeb92724c9cb0ef704f1e3de))
+
+## [1.8.2](https://github.com/VerifiedJoseph/gotify-api-php/releases/tag/v1.8.2) - 2024-06-24
+
+- Updated supported `gotify/server` version from 2.4.0 to 2.5.0. ([#385](https://github.com/VerifiedJoseph/gotify-api-php/pull/385), [`d236212`](https://github.com/VerifiedJoseph/gotify-api-php/commit/d236212d5c3e34fa4e48a28ca8460d4bfe0bf1e6))
+
 ## [1.8.1](https://github.com/VerifiedJoseph/gotify-api-php/releases/tag/v1.8.1) - 2023-12-04
 
 * Updated dependency `guzzlehttp/guzzle` from 7.8.0 to 7.8.1 ([#306](https://github.com/VerifiedJoseph/gotify-api-php/pull/306), [`0727221`](https://github.com/VerifiedJoseph/gotify-api-php/commit/07272215abc1a90de45dc2dec22120f48f6c3eeb))